move file to rich field folder.
