+++ packages/fields/src/fields/rich/index.ts
@@ -9,6 +9,7 @@ export type {
 } from './DrawingPad.js';
 export { DiagramField } from './DiagramField.js';
 export { DisplayField } from './DisplayField.js';
+export { FileField } from './FileField.js';
 export { HtmlField } from './HtmlField.js';
 export { ImageField } from './ImageField.js';
 export { SignatureField } from './SignatureField.js';
